#54 wire add/remove/complete and save flow on planning page (#63)
* #54 wire add/remove/complete and save flow on planning page

* #54 wire add/remove/complete/save on planning page; fix semester badge labels

* #54 fix

* #54 pr changes

* #54 better message

+++ backend/src/main/java/com/studyplanner/dto/StudyPlanResponse.java
@@ -9,7 +9,7 @@ import lombok.Builder;
 public record StudyPlanResponse(
     @NotNull UUID externalId,
     String name,
-    @NotNull Integer completedCredits,
+    @NotNull Double completedCredits,
     @NotNull LocalDateTime startDate,
     @NotNull CurriculumResponse curriculum,
     @NotNull LocalDateTime creationDate) {}
